requirements.txt
apache-airflow
airflow-provider-duckdb
apache-airflow-providers-cncf-kubernetes
duckdb
dbt-duckdb
pandas
requests
python-dotenv
dlt
sdv
rdt
python-dateutil